requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
chardet==4.0.0
colorama==0.4.4
discord.py==1.6.0
idna==3.1
multidict==5.1.0
Pillow==8.1.2
qrcode==6.1
six==1.15.0
typing-extensions==3.7.4.3
yarl==1.6.3
